made CRUD for posts, categories and tags

+++ resources/views/categories/index.blade.php
@@ -5,7 +5,8 @@
 @section('body')
     <h1>Categories</h1>
 
-    <a href="form.php">Add</a>
+    <a class="btn btn-primary" href="form.php" role="button">Add</a>
+    
     <table class="table">
         <thead>
         <tr>
@@ -17,15 +18,17 @@
         </thead>
         <tbody>
         @foreach($categories as $category)
-        <tr>
-            <th scope="row">{{ $category->id }}</th>
-            <td>{{ $category->title }}</td>
-            <td>{{ $category->slug }}</td>
-            <td>
-                <a href="form.php?id={{ $category->id }}">Edit</a> |
-                <a href="delete.php?id={{ $category->id }}">Delete</a>
-            </td>
-        </tr>
+            @if($category->title !== 'Uncategorized')
+                <tr>
+                    <th scope="row">{{ $category->id }}</th>
+                    <td>{{ $category->title }}</td>
+                    <td>{{ $category->slug }}</td>
+                    <td>
+                        <a href="form.php?id={{ $category->id }}">Edit</a> |
+                        <a href="delete.php?id={{ $category->id }}">Delete</a>
+                    </td>
+                </tr>
+            @endif
         @endforeach
         </tbody>
     </table>
